+++ VWOSdk/Internal/Extension/Extensions.cs
@@ -17,6 +17,11 @@
 #pragma warning restore 1587
 using System.Collections.Generic;
 using System.Net.Http;
+using System;
+using System.IO;
+using System.Runtime.Serialization;
+using System.Runtime.Serialization.Formatters.Binary;
+using Newtonsoft.Json;
 
 namespace VWOSdk
 {
@@ -35,6 +40,18 @@ namespace VWOSdk
         public static bool IsLogTypeEnabled(this LogLevel specifiedLogLevel, LogLevel logLevel)
         {
             return logLevel <= specifiedLogLevel;
+        }
+
+        public static T CloneJson<T>(this T source)
+        {
+          
+            // initialize inner objects individually
+            // for example in default constructor some list property initialized with some values,
+            // but in 'source' these items are cleaned -
+            // without ObjectCreationHandling.Replace default constructor values will be added to result
+            var deserializeSettings = new JsonSerializerSettings {ObjectCreationHandling = ObjectCreationHandling.Replace};
+
+            return JsonConvert.DeserializeObject<T>(JsonConvert.SerializeObject(source), deserializeSettings);
         }
     }
 }